refactor: experimet -> experiment
