+++ test/default_context_data_deserializer_test.dart
@@ -5,7 +5,7 @@ import 'package:absmartly_sdk/default_context_data_serializer.dart';
 import 'package:absmartly_sdk/json/context_data.dart';
 import 'package:absmartly_sdk/json/experiment.dart';
 import 'package:absmartly_sdk/json/experiment_variant.dart';
-import 'package:absmartly_sdk/json/experimet_application.dart';
+import 'package:absmartly_sdk/json/experiment_application.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
